+++ src/tooling/piral-cli-esbuild/set-path.js
@@ -3,6 +3,7 @@ function computePath() {
     throw new Error();
   } catch (t) {
     const e = ('' + t.stack).match(/(https?|file|ftp|chrome-extension|moz-extension):\/\/[^)\n]+/g);
+
     if (e) {
       return e[0].replace(/^((?:https?|file|ftp|chrome-extension|moz-extension):\/\/.+)\/[^\/]+$/, '$1') + '/';
     }
@@ -11,5 +12,4 @@ function computePath() {
   return '/';
 }
 
-// tslint:disable-next-line:variable-name
 export const __bundleUrl__ = computePath();