+++ src/tooling/piral-cli-esbuild/src/configs/common.ts
@@ -15,6 +15,7 @@ export function createCommonConfig(
     assetNames: contentHash ? '[name]-[hash]' : '[name]',
     chunkNames: contentHash ? '[name]-[hash]' : '[name]',
     publicPath: './',
+    charset: 'utf8',
     sourcemap,
     loader: {
       '.png': 'file',
@@ -33,6 +34,8 @@ export function createCommonConfig(
       'process.env.NODE_ENV': JSON.stringify(development ? 'development' : 'production'),
       'process.env.BUILD_PCKG_NAME': JSON.stringify(process.env.BUILD_PCKG_NAME),
       'process.env.BUILD_PCKG_VERSION': JSON.stringify(process.env.BUILD_PCKG_VERSION),
+      'process.env.PIRAL_CLI_VERSION': JSON.stringify(process.env.PIRAL_CLI_VERSION),
+      'process.env.BUILD_TIME_FULL': JSON.stringify(process.env.BUILD_TIME_FULL),
     },
     plugins: [sassPlugin(), codegenPlugin()],
     target: ['esnext'],