+++ src/tooling/piral-cli-esbuild/src/plugins/banner-plugin.ts
@@ -1,22 +1,28 @@
+import type { SharedDependency } from 'piral-cli';
 import type { PluginObj } from '@babel/core';
 import type { Statement } from '@babel/types';
 import template from '@babel/template';
 
 export interface PluginOptions {
   name: string;
-  deps: string;
+  importmap: Array<SharedDependency>;
   requireRef: string;
   cssFiles: Array<string>;
 }
 
-export default function babelPlugin(): PluginObj {
+export default function babelPlugin({ types }): PluginObj {
   const debug = process.env.NODE_ENV === 'development';
 
   return {
     visitor: {
       Program(path, state) {
-        const { name, deps, requireRef, cssFiles } = state.opts as PluginOptions;
-        path.addComment('leading', `@pilet v:2(${requireRef},${deps})`, true);
+        const { name, importmap, requireRef, cssFiles } = state.opts as PluginOptions;
+        const deps = importmap.reduce((obj, dep) => {
+          obj[dep.id] = dep.ref;
+          return obj;
+        }, {});
+
+        path.addComment('leading', `@pilet v:2(${requireRef},${JSON.stringify(deps)})`, true);
 
         if (cssFiles.length > 0) {
           const bundleUrl = `function(){try{throw new Error}catch(t){const e=(""+t.stack).match(/(https?|file|ftp|chrome-extension|moz-extension):\\/\\/[^)\\n]+/g);if(e)return e[0].replace(/^((?:https?|file|ftp|chrome-extension|moz-extension):\\/\\/.+)\\/[^\\/]+$/,"$1")+"/"}return"/"}`;
@@ -35,6 +41,22 @@ export default function babelPlugin(): PluginObj {
           ].join('\n');
           path.node.body.push(template.ast(`(function(){${stylesheet}})()`) as Statement);
         }
+      },
+      ImportDeclaration(path, state) {
+        const { importmap } = state.opts as PluginOptions;
+
+        path.traverse({
+          Literal(path) {
+            if (path.node.type === 'StringLiteral') {
+              const name = path.node.value;
+              const entry = importmap.find((m) => m.name === name);
+
+              if (entry) {
+                path.replaceWith(types.stringLiteral(entry.id));
+              }
+            }
+          },
+        });
       },
     },
   };